Extracting accessibiliy functions into a new helper file
