+++ app/app/javascript/utilities/accessibility.js
@@ -1,0 +1,21 @@
+export function toggleErrorIds(target, error, id) {
+  const describedbyList = target.getAttribute("aria-describedby")
+  if (error && describedbyList.includes(id)) return
+  let describeIds
+
+  if (error) {
+    describeIds = id + " " + describedbyList
+  } else {
+    describeIds = describedbyList.replace(id, "").trim()
+  }
+
+  target.setAttribute("aria-describedby", describeIds)
+}
+
+export function updateAriaLiveRegion(text) {
+  const liveRegion = document.getElementById("live-announcer")
+  liveRegion.replaceChildren()
+  if (text) {
+    liveRegion.textContent = text
+  }
+}
